+++ src/NEventStore/Conversion/EventUpconverterPipelineHook.cs
@@ -18,7 +18,6 @@ namespace NEventStore.Conversion
         /// <summary>
         /// Initializes a new instance of the EventUpconverterPipelineHook class.
         /// </summary>
-        /// <param name="converters"></param>
         /// <exception cref="ArgumentNullException"></exception>
         public EventUpconverterPipelineHook(IDictionary<Type, Func<object, object>> converters)
         {
@@ -66,7 +65,8 @@ namespace NEventStore.Conversion
 
         private object Convert(object source)
         {
-            if (!_converters.TryGetValue(source.GetType(), out Func<object, object> converter))
+            Type sourceType = source.GetType();
+            if (!_converters.TryGetValue(sourceType, out Func<object, object> converter))
             {
                 return source;
             }
@@ -74,7 +74,7 @@ namespace NEventStore.Conversion
             object target = converter(source);
             if (Logger.IsEnabled(LogLevel.Debug))
             {
-                Logger.LogDebug(Resources.ConvertingEvent, source.GetType(), target.GetType());
+                Logger.LogDebug(Resources.ConvertingEvent, sourceType, target.GetType());
             }
 
             return Convert(target);