+++ src/NEventStore/EventUpconverterWireup.cs
@@ -37,7 +37,7 @@ namespace NEventStore
                     _assembliesToScan.AddRange(GetAllAssemblies());
                 }
 
-                IDictionary<Type, Func<object, object>> converters = GetConverters(_assembliesToScan);
+                var converters = GetConverters(_assembliesToScan);
                 if (converters.Count > 0)
                 {
                     return new EventUpconverterPipelineHook(converters);