+++ src/NEventStore/ICommitEvents.cs
@@ -24,9 +24,9 @@ namespace NEventStore
         IEnumerable<ICommit> GetFrom(string bucketId, string streamId, int minRevision, int maxRevision);
 
         /// <summary>
-        ///     Writes the to-be-commited events provided to the underlying persistence mechanism.
+        ///     Writes the to-be-committed events provided to the underlying persistence mechanism.
         /// </summary>
-        /// <param name="attempt">The series of events and associated metadata to be commited.</param>
+        /// <param name="attempt">The series of events and associated metadata to be committed.</param>
         /// <exception cref="ConcurrencyException" />
         /// <exception cref="StorageException" />
         /// <exception cref="StorageUnavailableException" />