+++ src/NEventStore/Logging/LogFactory.cs
@@ -1,9 +1,8 @@
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Logging.Abstractions;
+
 namespace NEventStore.Logging
 {
-    using System;
-    using Microsoft.Extensions.Logging;
-    using Microsoft.Extensions.Logging.Abstractions;
-
     /// <summary>
     ///     Provides the ability to get a new instance of the configured logger.
     /// </summary>
@@ -14,8 +13,7 @@ namespace NEventStore.Logging
         /// </summary>
         static LogFactory()
         {
-            var logger = NullLogger.Instance;
-            BuildLogger = _ => logger;
+            BuildLogger = _ => NullLogger.Instance;
         }
 
         /// <summary>