+++ src/NEventStore/Serialization/BinarySerializer.cs
@@ -1,4 +1,3 @@
-using System.Runtime.Serialization;
 using System.Runtime.Serialization.Formatters.Binary;
 using Microsoft.Extensions.Logging;
 using NEventStore.Logging;
@@ -11,8 +10,8 @@ namespace NEventStore.Serialization
     [Obsolete("BinaryFormatter serialization is obsolete and should not be used. See https://aka.ms/binaryformatter for more information.")]
     public class BinarySerializer : ISerialize
     {
-        private static readonly ILogger Logger = LogFactory.BuildLogger(typeof (BinarySerializer));
-        private readonly IFormatter _formatter = new BinaryFormatter();
+        private static readonly ILogger Logger = LogFactory.BuildLogger(typeof(BinarySerializer));
+        private readonly BinaryFormatter _formatter = new();
 
         /// <inheritdoc/>
         public virtual void Serialize<T>(Stream output, T graph)
@@ -31,7 +30,7 @@ namespace NEventStore.Serialization
             {
                 Logger.LogTrace(Messages.DeserializingStream, typeof(T));
             }
-            return (T) _formatter.Deserialize(input);
+            return (T)_formatter.Deserialize(input);
         }
     }
 }